refactor: pool page
