+++ src/pages/Pools/constants/index.ts
@@ -1,0 +1,1 @@
+export const xOCH_PRICE = 0.4;
