+++ src/pages/Pools/hooks/useLiquidityPool.ts
@@ -1,37 +1,31 @@
-import { MulticallQueryClient } from '@oraichain/common-contracts-sdk';
+import { MulticallQueryClient } from '@oraichain/common-contracts-sdk/build/Multicall.client';
+import { network } from '@oraichain/oraidex-common/build/network';
 import { useQueryClient } from '@tanstack/react-query';
-import { ReactComponent as BackIcon } from 'assets/icons/ic_back.svg';
-import { network } from 'config/networks';
 import useConfigReducer from 'hooks/useConfigReducer';
 import useLoadTokens from 'hooks/useLoadTokens';
-import Content from 'layouts/Content';
-import React, { useCallback } from 'react';
+import { useCallback } from 'react';
 import { useDispatch, useSelector } from 'react-redux';
-import { useNavigate, useParams } from 'react-router-dom';
+import { useParams } from 'react-router';
 import { updateLpPools } from 'reducer/token';
-import { RootState } from 'store/configure';
 import { PoolInfoResponse } from 'types/pool';
-import styles from './PoolDetailV3.module.scss';
-import { Earning } from './components/Earning';
-import { MyPoolInfo } from './components/MyPoolInfo/MyPoolInfo';
-import { OverviewPool } from './components/OverviewPool';
-import { fetchLpPoolsFromContract, useGetPoolDetail, useGetPools } from './hookV3';
-import { useGetPairInfo } from './hooks/useGetPairInfo';
+import { fetchLpPoolsFromContract } from './useFetchLpPool';
+import { useGetPairInfo } from './useGetPairInfo';
+import { useGetPoolDetail } from './useGetPoolDetail';
+import { useGetPools } from './useGetPools';
+import { RootState } from 'store/configure';
 
-const PoolDetailV3: React.FC = () => {
-  let { poolUrl } = useParams();
-  const navigate = useNavigate();
+export const useLiquidityPool = (poolUrl: string) => {
   const dispatch = useDispatch();
+  const queryClient = useQueryClient();
   const [address] = useConfigReducer('address');
+  const pools = useGetPools();
   const lpPools = useSelector((state: RootState) => state.token.lpPools);
+  console.log('poolUrl', poolUrl);
   const poolDetailData = useGetPoolDetail({ pairDenoms: poolUrl });
   const lpTokenBalance = BigInt(poolDetailData.info ? lpPools[poolDetailData.info?.liquidityAddr]?.balance || '0' : 0);
   const loadTokenAmounts = useLoadTokens();
   const setCachedLpPools = (payload: LpPoolDetails) => dispatch(updateLpPools(payload));
-  const pools = useGetPools();
-  const poolDetail = useGetPoolDetail({ pairDenoms: poolUrl });
-  const { refetchPairAmountInfo, refetchLpTokenInfoData } = useGetPairInfo(poolDetail);
-  const queryClient = useQueryClient();
+  const { refetchPairAmountInfo, refetchLpTokenInfoData } = useGetPairInfo(poolDetailData);
 
   const refetchAllLpPools = useCallback(async () => {
     if (pools.length === 0) return;
@@ -64,24 +58,9 @@ const PoolDetailV3: React.FC = () => {
     [address, pools]
   );
 
-  return (
-    <Content nonBackground>
-      <div className={styles.pool_detail}>
-        <div
-          className={styles.back}
-          onClick={() => {
-            navigate(`/pools`);
-          }}
-        >
-          <BackIcon className={styles.backIcon} />
-          <span>Back to all pools</span>
-        </div>
-        <OverviewPool poolDetailData={poolDetailData} />
-        <Earning onLiquidityChange={onLiquidityChange} />
-        <MyPoolInfo myLpBalance={lpTokenBalance} onLiquidityChange={onLiquidityChange} />
-      </div>
-    </Content>
-  );
+  return {
+    onLiquidityChange,
+    poolDetailData,
+    lpTokenBalance
+  };
 };
-
-export default PoolDetailV3;
