+++ src/pages/UniversalSwap/Component/AssetsTab.tsx
@@ -1,31 +1,25 @@
-import WalletIcon from 'assets/icons/wallet-v3.svg';
+import { CW20_DECIMALS } from '@oraichain/oraidex-common/build/constant';
+import { getSubAmountDetails, toAmount, toDisplay } from '@oraichain/oraidex-common/build/helper';
+import { CoinIcon } from '@oraichain/oraidex-common/build/network';
+import { TokenItemType, tokenMap } from '@oraichain/oraidex-common/build/token';
+import { isMobile } from '@walletconnect/browser-utils';
 import StakeIcon from 'assets/icons/stake.svg';
-import styles from './AssetsTab.module.scss';
+import WalletIcon from 'assets/icons/wallet-v3.svg';
 import cn from 'classnames/bind';
-import { TableHeaderProps, Table } from 'components/Table';
-import { AssetInfoResponse } from 'types/swap';
+import { Table, TableHeaderProps } from 'components/Table';
+import ToggleSwitch from 'components/ToggleSwitch';
+import { flattenTokens } from 'config/bridgeTokens';
+import { tokensIcon } from 'config/chainInfos';
 import { useCoinGeckoPrices } from 'hooks/useCoingecko';
+import useConfigReducer from 'hooks/useConfigReducer';
 import { getTotalUsd, getUsd, toSumDisplay, toTotalDisplay } from 'libs/utils';
+import { formatDisplayUsdt, numberWithCommas, toFixedIfNecessary } from 'pages/Pools/helpers';
+import { useGetMyStake } from 'pages/Pools/hooks/useGetMyStake';
+import { FC, useState } from 'react';
 import { useSelector } from 'react-redux';
 import { RootState } from 'store/configure';
-import { isMobile } from '@walletconnect/browser-utils';
-import {
-  CW20_DECIMALS,
-  CoinIcon,
-  TokenItemType,
-  getSubAmountDetails,
-  toAmount,
-  toDisplay,
-  tokenMap
-} from '@oraichain/oraidex-common';
-import { FC, useState } from 'react';
-import { isSupportedNoPoolSwapEvm } from '@oraichain/oraidex-universal-swap';
-import { useGetMyStake } from 'pages/Pools/hookV3';
-import useConfigReducer from 'hooks/useConfigReducer';
-import ToggleSwitch from 'components/ToggleSwitch';
-import { tokensIcon } from 'config/chainInfos';
-import { flattenTokens } from 'config/bridgeTokens';
-import { formatDisplayUsdt, numberWithCommas, toFixedIfNecessary } from 'pages/Pools/helpers';
+import { AssetInfoResponse } from 'types/swap';
+import styles from './AssetsTab.module.scss';
 
 const cx = cn.bind(styles);
 
@@ -52,12 +46,12 @@ export const AssetsTab: FC<{ networkFilter: string }> = ({ networkFilter }) => {
     label?: string;
     balance?: number | string;
   }[] = [
-      {
-        src: WalletIcon,
-        label: 'Total balance',
-        balance: formatDisplayUsdt(totalUsd)
-      }
-    ];
+    {
+      src: WalletIcon,
+      label: 'Total balance',
+      balance: formatDisplayUsdt(totalUsd)
+    }
+  ];
 
   if (!networkFilter || networkFilter === 'Oraichain') {
     listAsset = [
@@ -94,7 +88,7 @@ export const AssetsTab: FC<{ networkFilter: string }> = ({ networkFilter }) => {
         usd += getUsd(subAmount, t, prices);
       }
       const value = toDisplay(amount.toString(), t.decimals) * prices[t.coinGeckoId] || 0;
-      const tokenIcon = tokensIcon.find(token => token.coinGeckoId === t.coinGeckoId);
+      const tokenIcon = tokensIcon.find((token) => token.coinGeckoId === t.coinGeckoId);
       return {
         asset: t.name,
         chain: t.org,
@@ -112,7 +106,7 @@ export const AssetsTab: FC<{ networkFilter: string }> = ({ networkFilter }) => {
   const headers: TableHeaderProps<AssetInfoResponse> = {
     assets: {
       name: 'ASSET',
-      accessor: data => (
+      accessor: (data) => (
         <div className={styles.assets}>
           <div className={styles.left}>
             {theme === 'light' ? (
@@ -134,14 +128,14 @@ export const AssetsTab: FC<{ networkFilter: string }> = ({ networkFilter }) => {
     price: {
       name: 'PRICE',
       width: '23%',
-      accessor: data => <div className={styles.price}>${Number(data.price.toFixed(6))}</div>,
+      accessor: (data) => <div className={styles.price}>${Number(data.price.toFixed(6))}</div>,
       align: 'left'
     },
     balance: {
       name: 'BALANCE',
       width: '23%',
       align: 'left',
-      accessor: data => (
+      accessor: (data) => (
         <div className={cx('balance', `${!data.balance && 'balance-low'}`)}>
           {numberWithCommas(toFixedIfNecessary(data.balance.toString(), isMobile() ? 3 : 6))}{' '}
           <span className={cx('balance-assets')}>{data.asset}</span>
@@ -153,7 +147,7 @@ export const AssetsTab: FC<{ networkFilter: string }> = ({ networkFilter }) => {
       width: '24%',
       align: 'left',
       padding: '0px 8px 0px 0px',
-      accessor: data => {
+      accessor: (data) => {
         return (
           <div className={styles.valuesColumn}>
             <div>
